Implementar ciclo completo de facturación y anulación
Se implementa la lógica de negocio completa para las facturas, cubriendo la creación, anulación, restitución de stock y notificaciones por correo electrónico, tanto en el backend como en el frontend.

**Backend:**
- Se implementa la generación automática de facturas (`Factura`) al momento de marcar un pedido (`Pedido`) como `ENTREGADO`.
- Se añade soporte para la facturación de pedidos pagados con `EFECTIVO` (a través de un nuevo endpoint para el cajero) y `MERCADO_PAGO` (disparado por el flujo de estados del pedido).
- Se crea la funcionalidad para anular facturas, la cual actualiza el estado de la factura a `ANULADA` y el del pedido asociado a `CANCELADO`.
- Se implementa la lógica de restitución de stock en `StockInsumoSucursal` al anular una factura, manejando correctamente tanto insumos de venta directa como los componentes de artículos manufacturados.
- Se integra un `EmailService` para enviar notificaciones al cliente:
    - Un correo con el detalle de la factura al momento de la creación.
    - Un correo de "Nota de Crédito" al momento de la anulación.
- Se corrigen los tipos de datos para los IDs de Mercado Pago (`Integer` a `Long`) para prevenir errores de formato numérico.

**Frontend:**
- Se añade la funcionalidad para que el cliente pueda visualizar sus facturas.
- Se crea el componente `FacturaDetailModal.tsx` para mostrar el detalle de una factura.
- Se modifica la página `MyOrdersPage.tsx` para que, al hacer clic en "Ver Factura", se obtengan los datos completos de la misma y se muestren en el nuevo modal.
- Se actualizan los tipos en `types.ts` (`PedidoResponse`) para reflejar los nuevos datos que envía el backend.

**Archivos Modificados/Agregados Clave:**
- **Backend:**
    - `services/FacturaServiceImpl.java`
    - `services/EmailServiceImpl.java` (nuevo)
    - `services/PedidoServiceImpl.java`
    - `controllers/FacturaController.java`
    - `controllers/PedidoController.java`
    - `entities/Factura.java`
    - `dtos/FacturaCreateRequestDTO.java`
    - `dtos/FacturaResponseDTO.java`
    - `dtos/PedidoResponseDTO.java`
- **Frontend:**
    - `pages/MyOrdersPage.tsx`
    - `components/facturas/FacturaDetailModal.tsx` (nuevo)
    - `types/types.ts`

+++ ElBuenSabor/src/main/java/com/powerRanger/ElBuenSabor/entities/Factura.java
@@ -33,8 +33,8 @@ public class Factura {
     private LocalDate fechaFacturacion;
 
     // Los campos de MercadoPago son opcionales
-    private Integer mpPaymentId;
-    private Integer mpMerchantOrderId;
+    private Long mpPaymentId;
+    private Long mpMerchantOrderId;
     private String mpPreferenceId;
     private String mpPaymentType;
 
@@ -74,10 +74,10 @@ public class Factura {
     public void setId(Integer id) { this.id = id; }
     public LocalDate getFechaFacturacion() { return fechaFacturacion; }
     public void setFechaFacturacion(LocalDate fechaFacturacion) { this.fechaFacturacion = fechaFacturacion; }
-    public Integer getMpPaymentId() { return mpPaymentId; }
-    public void setMpPaymentId(Integer mpPaymentId) { this.mpPaymentId = mpPaymentId; }
-    public Integer getMpMerchantOrderId() { return mpMerchantOrderId; }
-    public void setMpMerchantOrderId(Integer mpMerchantOrderId) { this.mpMerchantOrderId = mpMerchantOrderId; }
+    public Long getMpPaymentId() { return mpPaymentId; }
+    public void setMpPaymentId(Long mpPaymentId) { this.mpPaymentId = mpPaymentId; }
+    public Long getMpMerchantOrderId() { return mpMerchantOrderId; }
+    public void setMpMerchantOrderId(Long mpMerchantOrderId) { this.mpMerchantOrderId = mpMerchantOrderId; }
     public String getMpPreferenceId() { return mpPreferenceId; }
     public void setMpPreferenceId(String mpPreferenceId) { this.mpPreferenceId = mpPreferenceId; }
     public String getMpPaymentType() { return mpPaymentType; }
